nu virker det hele med smårettelser i modeller, repositories  og services ift metodekald

+++ src/main/java/org/example/adventurexp/model/Equipment.java
@@ -9,7 +9,7 @@ public class Equipment {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     private String name;
     private int totalSets, usableSets;
@@ -29,11 +29,11 @@ public class Equipment {
 
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
